Add tests for AllowAny()
